module 3 practice 1 fix

+++ module3/05/main.c
@@ -9,7 +9,7 @@
 #include <string.h>
 #include <signal.h>
 
-int file_locked = 0;
+volatile int file_locked = 0;
 
 void sigusr1_handler(int sig) {
     file_locked = 1;
@@ -74,7 +74,11 @@ int main(int argc, char *argv[]) {
             
 
             num = rand() % 1000;
-            write(pipefd[1], &num, sizeof(num));
+            if (write(pipefd[1], &num, sizeof(num)) != sizeof(num)) {
+                perror("write");
+                close(pipefd[1]);
+                exit(EXIT_FAILURE);
+            }
         }
 
         close(pipefd[1]);
@@ -85,18 +89,51 @@ int main(int argc, char *argv[]) {
 
         close(pipefd[1]);
         int output_fd = open("output.txt", O_WRONLY | O_CREAT | O_TRUNC, 0644);
+        if (output_fd == -1) {
+            perror("open");
+            close(pipefd[0]);
+            exit(EXIT_FAILURE);
+        }
 
         for (int i = 0; i < num_count; i++) {
             int num;
-            read(pipefd[0], &num, sizeof(num));
+            ssize_t bytes_read;
+            char *buffer = (char *)&num;
+            size_t bytes_remaining = sizeof(num);
+            while (bytes_remaining > 0) {
+                bytes_read = read(pipefd[0], buffer, bytes_remaining);
+                if (bytes_read == -1) {
+                    perror("read");
+                    close(pipefd[0]);
+                    close(output_fd);
+                    exit(EXIT_FAILURE);
+                } else if (bytes_read == 0) {
+                    fprintf(stderr, "Неожиданный конец передачи данных\n");
+                    close(pipefd[0]);
+                    close(output_fd);
+                    exit(EXIT_FAILURE);
+                }
+                bytes_remaining -= bytes_read;
+                buffer += bytes_read;
+            }
 
             char buf[32];
             int len = snprintf(buf, sizeof(buf), "%d\n", num);
-            write(STDOUT_FILENO, buf, len);
-            write(output_fd, buf, len);
+            if (write(STDOUT_FILENO, buf, len) != len) {
+                perror("write stdout");
+                close(pipefd[0]);
+                close(output_fd);
+                exit(EXIT_FAILURE);
+            }
+            if (write(output_fd, buf, len) != len) {
+                perror("write file");
+                close(pipefd[0]);
+                close(output_fd);
+                exit(EXIT_FAILURE);
+            }
 
             kill(pid, SIGUSR1);
-            FILE *file = fopen("input.txt", "w");
+            FILE *file = fopen("input.txt", "a");
             if (file == NULL) {
                 perror("Failed to open file");
                 exit(EXIT_FAILURE);
@@ -108,7 +145,13 @@ int main(int argc, char *argv[]) {
 
         close(pipefd[0]);
         close(output_fd);
-        wait(NULL);
+        int status;
+        wait(&status);
+        if (!WIFEXITED(status))
+        {
+            printf("child not exited");
+            exit(0);
+        }
     }
 
     return EXIT_SUCCESS;
